+++ resources/views/acts/edit.blade.php
@@ -26,14 +26,14 @@
                     </div>
 
                     <div class="flex items-center gap-4">
-                        <x-primary-button>{{ __('Save') }}</x-primary-button>
+                        <x-primary-button :icon="true">{{ __('Save') }}</x-primary-button>
                     </div>
                 </form>
 
                 <form method="POST" action="{{ route('acts.destroy', $act) }}" class="mt-6" onsubmit="return confirm('{{ __('Are you sure you want to delete this act?') }}')">
                     @csrf
                     @method('DELETE')
-                    <x-danger-button>{{ __('Delete Act') }}</x-danger-button>
+                    <x-danger-button :icon="true">{{ __('Delete Act') }}</x-danger-button>
                 </form>
             </x-card>
         </div>