+++ resources/views/admin/settings/edit.blade.php
@@ -66,7 +66,7 @@
                     </div>
 
                     <div class="flex items-center gap-4">
-                        <x-primary-button>{{ __('Save') }}</x-primary-button>
+                        <x-primary-button :icon="true">{{ __('Save') }}</x-primary-button>
 
                         <a
                             href="{{ route('robots.txt') }}"