+++ resources/views/chapters/edit.blade.php
@@ -36,14 +36,14 @@
                     </div>
 
                     <div class="flex items-center gap-4">
-                        <x-primary-button>{{ __('Save') }}</x-primary-button>
+                        <x-primary-button :icon="true">{{ __('Save') }}</x-primary-button>
                     </div>
                 </form>
 
                 <form method="POST" action="{{ route('chapters.destroy', $chapter) }}" class="mt-6" onsubmit="return confirm('{{ __('Are you sure you want to delete this chapter?') }}')">
                     @csrf
                     @method('DELETE')
-                    <x-danger-button>{{ __('Delete Chapter') }}</x-danger-button>
+                    <x-danger-button :icon="true">{{ __('Delete Chapter') }}</x-danger-button>
                 </form>
             </x-card>
         </div>