+++ resources/views/projects/edit.blade.php
@@ -97,14 +97,14 @@
                     </fieldset>
 
                     <div class="flex items-center gap-4">
-                        <x-primary-button>{{ __('Save') }}</x-primary-button>
+                        <x-primary-button :icon="true">{{ __('Save') }}</x-primary-button>
                     </div>
                 </form>
 
                 <form method="POST" action="{{ route('projects.destroy', $project) }}" class="mt-6" onsubmit="return confirm('{{ __('Are you sure you want to delete this project?') }}')">
                     @csrf
                     @method('DELETE')
-                    <x-danger-button>{{ __('Delete Project') }}</x-danger-button>
+                    <x-danger-button :icon="true">{{ __('Delete Project') }}</x-danger-button>
                 </form>
             </x-card>
         </div>